Detect that a parameter is mispelled.
Mispelled parameters used to be ignored. This update contains an explicit list of meta-parameters. Now, if you mispell a parameter (either "aws" meta-parameter or AWS parameter), a warning is printed, and execution is aborted.

Example:

$ aws register --name demo-ami --architecutre i386 --kernel aki-01f58a53 --ramdisk ari-25f58a77 -b /dev/sda=snap-3bb1db53:3:false
--architecutre: mispelled meta parameter?

More detail. A "meta-parameter" is a parameter that controls the "aws" program, whereas a parameter controls the AWS service. (For example, --xml versus --root-device-name.) The @cmd structure has always contained a list of all supported AWS parameters. Any parameter that isn't listed in @cmd was assumed to be a meta-parameter. If you mispelled a parameter (e.g., --architecutre), it was treated as a meta-parameter and ignored.

This update contains an explicit list of all meta-parameters. If a parameter isn't listed explicitely (either as a parameter or a meta-parameter), an error is displayed, and execution is aborted.

WARNING: It's possible that the new explicit list of meta-parameters is incomplete. If you use one of the meta-parameters that I missed, your properly constructed command will fail. If so, let me know what I missed, and I will add the missing meta-parameter to the list.