Feed curl config file to "curl" if one exists
First of all, thanks for the script. Really helpful, saved me a ton of time.

The patch checks if ~/.awscurlrc file exists and if it does then adds "-K ~/.awscurlrc" to the curl command line after "-q" (which as you know suppresses reading curl's config file from default locations).

This is needed for one particular case. I run a setup that does not have default directory with CA bundles. So running curl fails with "Can't check peer's certificate" and this is intentional. There's a directory with CA certs and all scripts that curl something off https:// explicitly specify this directory with --capath.

Since aws run curl with -q, I have no way to make it read from this CA directory. Hence the patch - I need a way for aws to run 'curl' with a custom config file, so that I could stick --capath there.